Merge pull request #3048 from MikeHolman:rejitexception

In some cases it's possible to have a rejit exception which wants to disable some optimization even if you don't have profile info. For this to work it required refactoring a bit in how we check if an optimization is disabled, and while I was here I decided to do a bit of cleanup around the RejitReason enum as well.

+++ lib/Common/Exceptions/RejitException.h
@@ -24,7 +24,7 @@ namespace Js
 
         const char *ReasonName() const
         {
-            return RejitReasonNames[reason];
+            return RejitReasonNames[static_cast<byte>(reason)];
         }
     };
 }
